Avoid using two copies of Aphrodite for no-important
After updating to v2 of this package, we noticed that our AMP builds
started failing with errors like

> Cannot automatically buffer without a document

After some investigation I believe I understand why this happened. The
main difference is that Aphrodite v2 is built using rollup, which means
that the default import and the no-important import are two entirely
separate self-contained modules. And since these modules maintain state,
it is important for them to be the same.

To fix this, I am moving the importing of `aphrodite` or
`aphrodite/no-important` up to the entry point level of this package,
and passing down the useful parts to be consumed by the factory. This
ensures that the same version of Aphrodite is used throughout.

This should have been caught by tests if we hadn't suppressed style
injection, which we don't need to do in this package because of how we
interact with Aphrodite, and if we were using the resolve method and
setting the AMP environment variable. I've addressed all of these issues
in this PR.

While testing this I noticed that when the AMP environment variable is
set, it never uses `!important` even if the non AMP styles use
`!important`. I think this is probably fine due to the constraints of
AMP, but I wasn't entirely sure if it was intentional. I added some
tests to showcase this behavior.

We should follow this up with work in Aphrodite itself that makes this
safer.