Add fallback for factory old behavior
In fixing the critical bug that shipped with v2, I needed to change the
API of this function, which makes the bugfix a semver major change. To
make this a semver-minor change, I am adding a fallback here so it still
works as it used to (in a buggy way) if this function is called directly
by some unknown outside consumer.

I think we could potentially detect whether to import from aphrodite or
aphrodite/no-important based on the shape of `aphroditeInterface` here,
but since I don't expect anyone to actually be hitting this codepath, I
didn't bother. The console warning should be enough to help folks
resolve this issue without too much work.